added task delete validation

+++ src/main/java/com/BESourceryAdmissionTool/task/services/TaskService.java
@@ -9,6 +9,8 @@ import com.BESourceryAdmissionTool.task.dto.FullTaskDto;
 import com.BESourceryAdmissionTool.task.dto.TaskDto;
 import com.BESourceryAdmissionTool.task.exceptions.TaskNameAlreadyExistsException;
 import com.BESourceryAdmissionTool.task.exceptions.TaskNotFoundException;
+import com.BESourceryAdmissionTool.task.exceptions.UserNotEqualTaskAuthorException;
+import com.BESourceryAdmissionTool.task.exceptions.UserNotLoggedInException;
 import com.BESourceryAdmissionTool.task.model.Task;
 import com.BESourceryAdmissionTool.task.repositories.TaskRepository;
 import com.BESourceryAdmissionTool.task.requests.TaskRequest;
@@ -18,6 +20,8 @@ import com.BESourceryAdmissionTool.user.model.User;
 import com.BESourceryAdmissionTool.user.repositories.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -62,6 +66,19 @@ public class TaskService {
         if (task.isEmpty()) {
             throw new TaskNotFoundException("Task not found");
         }
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String currentPrincipalEmail = authentication.getName();
+        String taskAuthorEmail = task.get().getAuthor().getEmail();
+
+        if(currentPrincipalEmail == null){
+            throw new UserNotLoggedInException("No User is logged in at the moment.");
+        }
+
+        if(!currentPrincipalEmail.equals(taskAuthorEmail) ){
+            throw new UserNotEqualTaskAuthorException("Can not delete task. You are not the author of the task.");
+        }
+
         answerRepository.deleteAnswersByTask(task.get());
         taskRepository.deleteById(id);
     }
